+++ src/table/components/default-cell-editors/aid-editor.tsx
@@ -3,19 +3,36 @@ import type { AIDFieldProps } from '../../../scalars/components/aid-field/types.
 import type { CellContext, DataType } from '../../table/types.js'
 
 export const buildAidCellEditor = <T extends DataType>(
-  aidFieldProps: Omit<
-    AIDFieldProps,
-    | 'name'
-    | 'value'
-    | 'onChange'
-    | 'autoComplete'
-    | 'fetchOptionsCallback'
-    | 'fetchSelectedOptionCallback'
-    | 'previewPlaceholder'
-  >
+  aidFieldProps: Omit<AIDFieldProps, 'name' | 'value' | 'onChange'>
 ) => {
   const AidCellEditor = (value: unknown, onChange: (newValue: unknown) => void, context: CellContext<T>) => {
-    const aidValue = typeof value === 'string' ? value : String(value ?? '')
+    const aidValue = typeof value === 'string' ? value : (value?.toString() ?? '')
+
+    if (aidFieldProps.autoComplete !== false && aidFieldProps.fetchOptionsCallback) {
+      const { autoComplete: _, fetchOptionsCallback, ...restProps } = aidFieldProps
+
+      return (
+        <AIDField
+          name={context.column.field}
+          className="max-w-full"
+          autoFocus
+          fetchOptionsCallback={fetchOptionsCallback}
+          {...restProps}
+          value={aidValue}
+          onChange={(value: string) => {
+            onChange(value)
+          }}
+        />
+      )
+    }
+
+    const {
+      autoComplete: _1,
+      fetchOptionsCallback: _2,
+      fetchSelectedOptionCallback: _3,
+      previewPlaceholder: _4,
+      ...restProps
+    } = aidFieldProps
 
     return (
       <AIDField
@@ -23,7 +40,7 @@ export const buildAidCellEditor = <T extends DataType>(
         className="max-w-full"
         autoFocus
         autoComplete={false}
-        {...aidFieldProps}
+        {...restProps}
         value={aidValue}
         onChange={(value: string) => {
           onChange(value)