+++ src/table/logic/table-api.ts
@@ -469,10 +469,7 @@ class TableApi<TData> implements PrivateTableApiBase<TData> {
         })
       }
 
-      this._getState().dispatch?.({
-        type: 'SELECT_CELL',
-        payload: { row: selectedCell.row, column: selectedCell.column },
-      })
+      this.selection.selectCell(selectedCell.row, selectedCell.column)
     }
 
     // Always trigger exit event at the end